Move GitHub and Zulip API clients to a separate module
